requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
backoff==2.2.1
certifi==2025.6.15
charset-normalizer==3.4.2
colorlog==6.9.0
datasets==2.14.7
dill==0.3.7
distro==1.9.0
dspy-ai==2.4.3
exceptiongroup==1.3.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2023.10.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.1
idna==3.10
jiter==0.10.0
joblib==1.3.2
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.5.1
multiprocess==0.70.15
numpy==2.0.2
openai==1.91.0
optuna==4.4.0
packaging==25.0
pandas==2.3.0
propcache==0.3.2
pyarrow==20.0.0
pyarrow-hotfix==0.7
pydantic==2.7.0
pydantic_core==2.18.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
structlog==25.4.0
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
xxhash==3.5.0
yarl==1.20.1
